Add tests for FP1 functionality
