+++ tests/test_sensors.py
@@ -584,134 +584,6 @@ async def test_power_sensor(input, expected, deconz_sensor):
         assert getattr(sensor, attr) == value
 
 
-async def test_configure_presence_sensor(
-    mock_aioresponse, deconz_session, deconz_called_with
-):
-    """Verify that configuring presence sensor works."""
-    presence = deconz_session.sensors.presence
-
-    mock_aioresponse.put("http://host:80/api/apikey/sensors/0/config")
-    await presence.set_config("0", delay=10, duration=20, sensitivity=1)
-    assert deconz_called_with(
-        "put",
-        path="/sensors/0/config",
-        json={"delay": 10, "duration": 20, "sensitivity": 1},
-    )
-
-    mock_aioresponse.put("http://host:80/api/apikey/sensors/0/config")
-    await presence.set_config("0", delay=1)
-    assert deconz_called_with(
-        "put",
-        path="/sensors/0/config",
-        json={"delay": 1},
-    )
-
-    mock_aioresponse.put("http://host:80/api/apikey/sensors/0/config")
-    await presence.set_config("0", duration=2)
-    assert deconz_called_with(
-        "put",
-        path="/sensors/0/config",
-        json={"duration": 2},
-    )
-
-    mock_aioresponse.put("http://host:80/api/apikey/sensors/0/config")
-    await presence.set_config("0", sensitivity=3)
-    assert deconz_called_with(
-        "put",
-        path="/sensors/0/config",
-        json={"sensitivity": 3},
-    )
-
-
-async def test_presence_sensor(mock_aioresponse, deconz_sensor, deconz_called_with):
-    """Verify that presence sensor works."""
-    sensor = await deconz_sensor(
-        {
-            "config": {
-                "alert": "none",
-                "battery": 100,
-                "delay": 0,
-                "ledindication": False,
-                "on": True,
-                "pending": [],
-                "reachable": True,
-                "sensitivity": 1,
-                "sensitivitymax": 2,
-                "usertest": False,
-            },
-            "ep": 2,
-            "etag": "5cfb81765e86aa53ace427cfd52c6d52",
-            "manufacturername": "Philips",
-            "modelid": "SML001",
-            "name": "Motion sensor 4",
-            "state": {"lastupdated": "2019-05-05T14:37:06", "presence": False},
-            "swversion": "6.1.0.18912",
-            "type": "ZHAPresence",
-            "uniqueid": "00:17:88:01:03:28:8c:9b-02-0406",
-        },
-    )
-
-    assert sensor.ZHATYPE == ("ZHAPresence", "CLIPPresence")
-
-    assert sensor.presence is False
-    assert sensor.dark is None
-    assert sensor.delay == 0
-    assert sensor.duration is None
-    assert sensor.sensitivity == 1
-    assert sensor.max_sensitivity == 2
-
-    # DeconzSensor
-    assert sensor.battery == 100
-    assert sensor.ep == 2
-    assert sensor.low_battery is None
-    assert sensor.on is True
-    assert sensor.reachable is True
-    assert sensor.tampered is None
-    assert sensor.secondary_temperature is None
-
-    # DeconzDevice
-    assert sensor.deconz_id == "/sensors/0"
-    assert sensor.etag == "5cfb81765e86aa53ace427cfd52c6d52"
-    assert sensor.manufacturer == "Philips"
-    assert sensor.model_id == "SML001"
-    assert sensor.name == "Motion sensor 4"
-    assert sensor.software_version == "6.1.0.18912"
-    assert sensor.type == "ZHAPresence"
-    assert sensor.unique_id == "00:17:88:01:03:28:8c:9b-02-0406"
-
-    mock_aioresponse.put("http://host:80/api/apikey/sensors/0/config")
-    await sensor.set_config(delay=10, duration=20, sensitivity=1)
-    assert deconz_called_with(
-        "put",
-        path="/sensors/0/config",
-        json={"delay": 10, "duration": 20, "sensitivity": 1},
-    )
-
-    mock_aioresponse.put("http://host:80/api/apikey/sensors/0/config")
-    await sensor.set_config(delay=1)
-    assert deconz_called_with(
-        "put",
-        path="/sensors/0/config",
-        json={"delay": 1},
-    )
-
-    mock_aioresponse.put("http://host:80/api/apikey/sensors/0/config")
-    await sensor.set_config(duration=2)
-    assert deconz_called_with(
-        "put",
-        path="/sensors/0/config",
-        json={"duration": 2},
-    )
-
-    mock_aioresponse.put("http://host:80/api/apikey/sensors/0/config")
-    await sensor.set_config(sensitivity=3)
-    assert deconz_called_with(
-        "put",
-        path="/sensors/0/config",
-        json={"sensitivity": 3},
-    )
-
-
 async def test_pressure_sensor(deconz_sensor):
     """Verify that pressure sensor works."""
     sensor = await deconz_sensor(
